modified:   os_version/Application/Hardware/Inc/esp-01s.h 	modified:   os_version/Application/Hardware/Src/esp-01s.c

+++ os_version/Application/Hardware/Inc/esp-01s.h
@@ -4,7 +4,7 @@
 #include "stm32f4xx.h"
 
 void Usart2_SendByte(uint8_t);
-uint16_t Usart2_ReceiveByte(void);
+uint8_t Usart2_ReceiveByte(void);
 
 void ESP_Init(void);
 void ESP_GpioInit(void);
